requirements.txt
ansi2html==1.8.0
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
cryptography==41.0.4
dash==2.13.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.0
firebase-admin==6.2.0
firestore==0.0.8
Flask==2.2.5
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.12.3
iso8601==2.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msgpack==1.0.7
nest-asyncio==1.5.8
numpy==1.24.4
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
plotly==5.17.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
retrying==1.3.4
rsa==4.9
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
tornado==6.3.3
traitlets==5.10.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.8
Werkzeug==2.2.3
zipp==3.17.0
